ref(transport): separate legacy and mock files
Rename the legacy files and extract the mock before changing their implementations.
